library_management/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.18
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
